refactor(ci): split the Sentry CI-coverage checker into a core module
The mutation probes added by the two commits before this one only work
because the predicates take their input as an argument. Keeping them in
the same file as the reads they used to close over left the entry point
at ~1,100 lines and mixed two kinds of code: pure judgement, and the
file/process access that makes the judgement about this repo.

scripts/check-sentry-suites-in-ci-core.mjs now holds the shell grammar,
the workflow predicates, and alias resolution — no file reads, no
execFileSync, every input a parameter. The entry point keeps the reads,
the three external-process probes, the repo-specific policy constants,
and every `test()`, unchanged in name and order.

Wiring the split so it cannot become its own blind spot: the new module
joins the gate's Sentry routing case with a `run_gate` assertion, and the
entry point asserts (through V8's parser, not a substring search) that it
still imports the core module. `scripts/**/*.mjs` already routes the new
file in CI, and the ci.yml step is unchanged.
